wip - astro - refine pagination, auth methods
