+++ apps/web-astro/src/pages/app/episodes/index.astro
@@ -1,14 +1,22 @@
 ---
 import { EpisodeList, FORM_ACTIONS, FORM_SUBJECTS } from '@pdcstrrss/ui';
-import Layout from 'apps/web-astro/src/layouts/Layout.astro';
+import { LOGIN_API_URL } from '../../../lib/useAuth';
+import Layout from '../../../layouts/Layout.astro';
 import { getEpisodes, updateEpisodes } from '../../../lib/useEpisodes.server';
-import usePaginationFilters from 'apps/web-astro/src/lib/usePagination';
+import { usePaginationFilters } from '../../../lib/usePagination';
+import { getUserFromRequest } from '../../../lib/useUser';
 
-await updateEpisodes(Astro);
-const data = await getEpisodes(Astro);
+const { request } = Astro;
+const { url, method } = request;
+
+const user = await getUserFromRequest(Astro);
+if (!user) return Astro.redirect(LOGIN_API_URL);
+
+if (method === 'POST') await updateEpisodes({ request, user });
+const data = await getEpisodes({ url, user });
 const { Pagination, itemsPerPage, currentPage, urlTransformer } = usePaginationFilters({
   ...data,
-  url: Astro.request.url,
+  url,
 });
 ---
